+++ src/main/java/com/finkeflo/cpi/kafka/ProducerBatchHelper.java
@@ -215,6 +215,7 @@ public final class ProducerBatchHelper {
         message.setHeader("SAP_Receiver", topic);
         message.setHeader("CamelKafkaTopic", topic);
         message.setHeader("CpiKafkaPlusTopic", topic);
+        message.setHeader("CpiKafkaPlusStatus", "OK");
         message.setHeader("CpiKafkaPlusRecordCount", result.getRecordCount());
         message.setHeader("CpiKafkaPlusBatchInputFormat", batchMode);
         message.setHeader("CpiKafkaPlusFirstOffset", result.getFirstOffset());