+++ src/test/java/com/finkeflo/cpi/kafka/ProducerBatchSendIT.java
@@ -290,6 +290,7 @@ public class ProducerBatchSendIT {
         Assert.assertEquals(topic, out.getHeader("SAP_Receiver"));
         Assert.assertEquals(topic, out.getHeader("CamelKafkaTopic"));
         Assert.assertEquals(topic, out.getHeader("CpiKafkaPlusTopic"));
+        Assert.assertEquals("OK", out.getHeader("CpiKafkaPlusStatus"));
         Assert.assertEquals(3, out.getHeader("CpiKafkaPlusRecordCount"));
         Assert.assertEquals("JSON_ARRAY", out.getHeader("CpiKafkaPlusBatchInputFormat"));
         Assert.assertNotNull(out.getHeader("CpiKafkaPlusFirstOffset"));